:bug: Make window resizable
Two notable changes have been made here. First, the window no longer
forces itself to be fullscreen, since most people want to be able to
play in a window or use another size. Second, the scaling of the
game has changed to always force a 16:10 aspect ratio, which works
well on Macs and other devices. This also fixes some glitches in
resizing the window with the expand scale mode on the main menu.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>