:bug: Reverse direction of the FOV slider
The direction was flipped in such a way where lowering the slider
would zoom in, and raising the slider would zoom out. This was set
initially due to the constraints of the camera's field of view
property.

To fix this issue, the min and max values for the slider are set
to a negative value and flipped. When the value is changed and the
signal to update the FOV is emitted, the emitter will send over the
absolute value of the requested FOV instead of the original value,
thus preserving the FOV and allowing the direction to be reserved.

Additionally, the shortcuts to zoom in and out have been changed to
just + and - (keyboard-wise, this would mean = and -, since to use
+, you need to press Shift+=).

Signed-off-by: Marquis Kurt <software@marquiskurt.net>