:lipstick: Increase size of some UI elements
The on-screen controls were difficult to control, so the scale has
been adjusted to accomodate for different sizes. The toolbar also
has been upsized slightly.

The "quit" option should also now be hidden on iOS, Android, and
the web, since it's pointless to show this button on these platforms.

The aspect ratio has also been updated to use 16:9 instead of 16:10.
This may be adjusted per-platform to keep things consistent on those
platforms. To accomodate for any possible size changes in the future,
the main menu world has been expanded a bit with more content.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>