:lipstick: Change Z-index of on-screen controls
The on-screen controls sat on top of the toolbars and tutorials,
which made it impossible to click any of the existing buttons in
said toolbars. The controls now sit underneath everything else,
which should make everything accessible again.

The positions of these controls have been slightly updated so that
it feels more natural to use.

It is important to note that, for the time being, these controls
are merely offered as an additional means of control for devices
with touchscreens, such as Windows 2-in-1 laptops, Macs using the
Sidecar feature on an iPad, and Linux devices with a touchscreen.

Porting the game over to mobile devices running iOS or Android has
not been confirmed at this stage, and it is unknown if the game will
be compatible on these platforms.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>