:children_crossing: Add lighting abilities to input devices
Previously, there were no visual indicators to determine whether
input devices were being activated or deactivated, though audio
cues are present to help mitigate this. This can be problematic
for players who may have audio impairments and are able to pick
up on those cues.

To resolve this issue, input devices now generate a light when
entering the scene tree. If LIGHT_WHEN_ACTIVE is turned on for
that input device, the light will be emitted according to the
other light parameters specified. All of the levels have been
updated to make use of this, with different device types emitting
lights of different colors.

This is just a small improvement which should help provide some
visual clarity on what's going on in the environment. Future commits
will focus on improving readability by adding visual indicators that
help identify what some input devices turn on, as well as a visual
indication of an exit door being opened.

Thanks to @andy-k for bringing this to my attention.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>