:recycle: Refactor exit door system
Previously, the exit door maintained a static appearance, and it
wasn't obvious when it was turned on and if it was the exit to
begin with.

Exit doors now take on a similar approach to the door node that is
used regularly for some of the levels. They emit their own light
when deactivated, and they visually appear different in-game. Due to
scaling to fit the bounds of the textures, it may appear off with
respect to scaling with everything else. This will need to be
addressed in a future version.

They function the same way as the previous system, with some minor
cosmetic changes when activated to indicate that the door is open.

Thanks to @andy-k for bringing this to my attention.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>