:bug: Resolve multiple issues
An issue where exit doors could not be opened when wearing the bird
costume has been resolved. The collision mask for detection was set
incorrectly on these doors.

An issue where some levers could not be used when wearing the bird
costume has been fixed, similar to the exit door cases.

The lever visual in tschmb_07 has been changed to an alarm clock
to maintain consistency with other timer-based inputs.

Attempts to solve an issue where carpet (collision-based) inputs
would be incorrectly triggered, or toggled on and off inconsistently.
The methods that detect when a player or object enters/exits the
collision field didn't check that the input was already in the
_entered array, and the removing when exiting only removed the first
item, or failed (proper method was erase, not remove). These checks
and changes have been added in hopes of fixing these input types.

Thanks to @andy-k for bringing these bugs forward.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>