:children_crossing: Replace second computer type with a server
In the previous prototypes, it was unclear that the second type of
computer, the one with an orange glow, requires the wearing of the
Swift bird costume.

To make this distinction clearer for those who are unable to discern
the new input type based upon the glow, the second computer type has
been replaced with a tall server rack.

Thanks to @andy-k for bringing this issue to my attention.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>