:children_crossing: Add barriers around levels
In the original game, and previous iterations, the player could go
far out of bounds when wearing the Swift costume. Although fun, the
player may try to solve puzzles by going out of bounds and entering
another room. There are other measures in place to keep this in check,
such as certain levels disabling the costume outright.

The barrier is another tilemap with collision layers and masks set
to listen for the Swift bird mask and all others, preventing the
player from going OOB. This should also help motivate the player
to interact with the exit door when it is open, preventing any
possible issues in the future.

Thanks to @andy-k for bringing this to my attention.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>