:bug: Resolve issues in collision-based inputs
In the previous prototypes, there were multiple issues with inputs
that relied on collision information to detect when it should be
activated. Commit 1f374b5 tried to resolve these issues with little
success by changing the methods called on the _entered array so that
the detection was more consistent.

First, the logic applied in 1f374b5 needed to also be applied to
Area2D nodes, since they are also included in the collision
detection system.

In _on_body_exited, the case for when the input was active and the
list of entered nodes was empty was not being reached due to a typo
on line 199 (it was checking for the list being of length -1 instead
of 0).

Given that the USB clone acts like a movable object but isn't a static
body, the Clone node has been changed to type Area2D to be consistent.
The methods for body and area detection have also been changed with this
change in mind.

Finally, carpets will now use the default ACTIVE_LIGHT_ENERGY value,
instead of the custom value of 0.3. Though it looks odd from a design
and visual perspective, this should make it clearer for players that
were unable to see the lighting effects before.

Thanks to @andy-k for bringing these issues to my attention.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>