:children_crossing: Improve movable object experience
Although barely functional, the current movable object system is
rough and conflicts with the player's player hint indicator too
much to be reliable.

To fix this, movable objects will have their own means of indicating
if it can be picked up by emitting a blue glow and tinting the object
blue. This drops the need to display a player hint or conflict with
other hints when being activated.

The player now includes a toggle_item_detection method which will
set whether a player is near an item, which will be accounted for
in input handling.

To resolve some minor issues with movable objects to inputs that
use the interact key or require both collisions and keypresses, the
input system has been refactored again. A new method
_update_active_checks will be called when a body or area enters or
exits the input's collider. This method now properly checks for the
different input types in a more robust manner than before.

Thanks to @andy-k for bringing the issues to my attention.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>