:bento: Include Aseprite source files for sprites
The README mentioned Aseprite source files for the character sprites
used in the game, but none were present. With the exception to the
main menu sprite, all were copied over from an older prototype of
a different game that has been recently archived (codename Rhemflash).

These sprites are to be licensed under the Creative Commons Attribution-
ShareAlike 4.0 International license.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>