:sparkles: Add pause menu
There is now a pause menu in place that can be access by pressing
the Escape key, the Start button on some controllers (aka Options
or + on some Nintendo controllers), or any variants. The pause
menu contains three basic options: resume the game, show the user
settings, or go back to the main menu.

In the future, if saving and loading is implemented for this game,
that option will also reside here.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>