:lipstick: Add fadeout transition to universe when switching levels
Rather than an abrupt ending to switch to the next level, the door
will now interpolate the modulate for the scene's root to fade out
the scene and then switch to the next level. This should make it
clear that the next level is starting.

In the future, the universe should play an alarm clock sound at the
beginning to indicate that the player is waking up, again and again.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>