:children_crossing: Implement visual indicators on inputs, steps
In previous iterations, inputs were not marked in a way that gave
a visual indication of what it activated. Similarly, output devices
did not have any visual indicators on what activated them.

This commit implements the signage necessary to indicate what inputs
activate what devices. In some stages, steps are labeled, like in
tschmb_06.

The signage available in the decor tileset will be deprecated in
favor of the signage tileset, which should be placed on top of
decor tilemaps.

^CMR-1 stage Done work Development 31m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>