:triangular_flag_on_post: Add debug setting for animations
In commit c1a082a (and per CMR-2), a new fadeout animation was
introduced when transitioning between levels to give a sense of
clarity while also making it obvious to the player that they are
waking up each time repeatedly.

The animation, however, is a bit long and can take a while during
testing and debugging. This commit adds a debug flag to toggle this
animation.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>