:construction: Implement SaveState class
The SaveState class is an abstract data class that represents a
suspended scene state. It contains necessary information such as
the time it was saved, the player's information, current level,
and the input states of the level, including activity and timer
durations.

The universe class has been updated so that it can resume from a
suspended state, as well as creating a SaveState with the new
Universe._suspend_to_state method.

Implementing the file saving and loading should come next, though
the groundwork is in place with the aforementioned method
Universe._resume_from_suspended_state.

^CMR-3 stage Develop work Development 42m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>