:sparkles: Fully implement save system
A new SaveUtils class has been introduced to load and save data to
a single save file (currently), along with encrypting the data.

Some modifications needed to be made to the save system, since it
appears that object storing in Godot wasn't working as intended.
For save states, a serialized JSON version is now stored in these
files.

In the universe, when setting up the world, the level will check
for a save file and resume the state if the level matches. This is
not a foolproof check, but this should work in most cases. Duration
of time-based inputs doesn't seem to be able to be restored well
because Timer.time_left is a readonly property. Further improvements
to this will need to be made.

The pause menu contains a button to save the game, and a notification
will appear in the top right on the pause menu to indicate if it was
successful. Translation files have been updated to accommodate for
this change.

^CMR-3 stage Review work Development 110m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>