:children_crossing: Add ability to switch bwetween languages
Given that the game supports locales and translations, having a
preferred locale setting is crucial. This commit adds the ability
to set the preferred locale in UserDefaults via the settings pane.

When set, the translation server will automatically set the correct
locale and update the text as needed. The settings pane will load in
all locales that are present in the translations file at runtime.

Additionally, an issue where the player's costume wasn't being loaded
correctly from a save has been resolved by calling change_costume
instead of changing the variable directly.

^CMR-7 stage Done work Development 1m
^CMR-5 stage Done work Development 29m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>