:sparkles: Implement audio buses for volume control
Audio buses have been implemented to handle the volume preferences
that were implemented in the previous commit. Audio streams will
use the SFX and Music audio buses to adjust their volume DB, which
adjusts their volume in-game. This resolves a few issues and the
feedback is instantaneous.

With these changes in place, UserDefaults.allow_music has been
deprecated and will be removed in a future release. If players want
to mute the music, they can adjust the volume sliders..

^CMR-11 work Development 49m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>