:children_crossing: Update level layouts
Using the tileset from the previous commit, the levels have been
slightly modified to follow the guidelines in CMR-15. This is
still a work in progress, so more may need to be done to make it
more distinctive. As such, the tileset may need some other visual
updates to fix pixel inconsistencies.

^CMR-15 stage Review work Development 25m
^CMR-16 stage Review work Artwork 5m
^CMR-17 stage Review work Development 10m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>