:children_crossing: Redesign tschmb_02
This level has been redesigned to show the concepts present more
intuitively.

There now exists two doors; one which leads to the exit door, and
the exit door itself. The first door is linked by two levers in the
main area, whereas the exit door has a direct connection to the
computer.

This should help demonstrate the two concepts introduced in a more
intuitive manner, showing the player that some doors are locked with
multiple inputs, as well as the computer being a viable input.

^CMR-17 work Development 10m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>