:children_crossing: Improve costume switching on controllers
In the past, the controller would attempt to switch costumes in a
linear fashion, increasing or decreasing the float value associated
with that costume. Although functional, it wasn't clear to the
player that this was happening, so pressing the bumper buttons to
switch between costumes was not responsive to the feedback at times.

To fix this, the list of allowed costumes that the HUD receives is
copied. When the player uses the bumper buttons, the list is popped
and pushed with the costume in the list, depending on which direction
to cycle through. This works similarly to how the original game handled
this feedback.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>