:feature: Add debugging page to settings window
Some of the settings only available via the debug feature setting
have been added to a new Debugging tab in the game's settings
window. These features are enabled when debug mode is, and they
will affect the environment like the original keyboard shortcuts
do (which are still in place).

The following settings have been added to the debug section of the
UserDefaults:

- level_name_in_toolbar (bool)
- fullbright (bool)

^CMR-22 stage Done work Development 49m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>