:sparkles: Add dialogue system to game
The dialogue system (ported over from an old prototype of a different
project) will allow text to be displayed via an output. This will be
used to help tell a bit more story.

To do this, the output specifies which dialogue set to load. This set
is pulled from the JSON file, which contains the keys for which to
display text. The UI will automatically pick the appropriate text based
on locale.

^CMR-12 stage Develop work Development 10m
^CMR-21 stage Develop work Development 10m
^CMR-14 stage Develop work Development 10m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>