:lipstick: Update design of dialogue box
In the previous beta, the dialogue system was introduced in an
attempt to make the gameplay more fun and easy to explain per
CMR-21. This seemed to be successful, so it was noted that the
design should be updated to make it feel less tacked on.

Dialogue boxes like what is provided are usually simple, so most
of the changes in this commit just amp the padding and show the
player's head in the USB costume.

This commit also updates the collision shapes in the arc_tileset
to accommodate for the new player collision shape introduced in
the previous beta.

^CMR-24 stage Staging work Development 18m
^CMR-21 work Development 18m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>