:children_crossing: Display some controller UI for controller input
Controller support exists for the game, but it's not immediately
clear what the control scheme is. This commit adds some UI elements
in the HUD that display the controls for switching between costumes
and restarting the level.

A future commit will add controller UI for movement and interaction.
How this will work is to be determined.

^CMR-18 work Artwork 5m
^CMR-18 work Development 15m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>