:children_crossing: Change focus modes on main menu buttons
In the previous betas, controllers could not select main menu buttons
because the focus modes were not set to describe the behavior in which
they operated.

This commit updates the Control.focus_mode on each button specifying
its neighbors so that controllers can select using the D-Pad more
efficiently. Additionally, if a controller is detected on startup,
the MENU_START button will automatically be focused.

This commit also deletes sensitive Godot files as mentioned in their
documentation.

^CMR-18 stage Test work Development 22m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>