:construction: Add snapcraft files
This commit introduces some files needed to create the snap package
so that it can be distributed on Itch.io or the Snap Store. It is
undecided whether this will be distributed through the Snap Store
or through Itch.io exclusively at this time due to limitations with
paid snaps.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>