:iphone: Resolve interaction issues with mobile interface
In the previous iterations, the interact and clone buttons did not
perform their respective actions correctly or did so inconsistently.
This was caused by the player's input handler only listening on a
keypress and not any given input.

This commit resolves the issue by changing the input method to
_unhandled_input instead of _unhandled_key_input. Additionally, this
commit resolves a "return value discarded" warning on the joypad,
though the return value isn't necessary.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>