:boom: Change rendering backend to GLES3
The game will now use GLES3 as its rendering backend, instead of
GLES2. This resolves some visual glitches on iOS and may improve
graphics on other devices.

To ensure compatibility is mostly preserved, the game will fall
back to GLES2. This should not change the game dramatically, since
the game was designed in GLES2 first.

This commit also changes the HUD to hide the walk tutorial when
on-screen controls are present. This should mitigate usage on mobile
devices and devices where on-screen controls are present by default.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>