:recycle: Make SaveUtils a singleton
The SaveUtils class is now a singleton, making all of its methods
static. Scripts that rely on SaveUtils to load and save suspended
states now call the singleton's classes as if they are static
methods.

In some methods that try to call for the Universe by navigating to
the tree and looking for its children, the index has been adjusted
to account for the SaveUtils singleton.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>