:iphone: Work on mobile interface design
The entirety of the dialogue box is now clickable when
UserDefaults.show_mobile_controls is enabled or when the device is
a mobile device. This should make tapping to advance easier.
Additionally, this will also hide the "Next" button commonly used on
the desktop.

The mobile HUD has been updated to show the pause button in the top
right corner of the screen, rather than the bottom center.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>